projects/misc/2026-01-26/1769393294746/requirements.txt
# Claude Agent SDK for Python
claude-agent-sdk>=0.1.0

# Environment variable management
python-dotenv>=1.0.0
